refactor health endpoints
